==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests

def download_one_file_of_raw_data(year: int, month: int) -> Path:
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL)

    if response.status_code == 200:
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'
        open(path, "wb").write(response.content)
        return path
    else:
        raise Exception (f'{URL} is not available')

download_one_file_of_raw_data(2022, 1)

import pandas as pd

rides = pd.read_parquet("../data/raw/rides_2022-01.parquet")
rides.head(20)

rides = rides[['tpep_pickup_datetime', 'PULocationID']]

rides = rides.rename(columns = {'tpep_pickup_datetime':'pickup_datetime', 'PULocationID':'pickup_location_id'})
rides.head(20)

rides['month'] = rides['pickup_datetime'].dt.month
rides.head()

rides

rides['pickup_datetime'].describe()

def correct_year_month(df, year, month):
    df['month'] = df['pickup_datetime'].dt.month
    df['year'] = df['pickup_datetime'].dt.year
    df_filter = df[(df['year'] == year) & (df['month'] == month)]
    df_filter = df_filter[['pickup_datetime', 'pickup_location_id']]
    return df_filter

rides = correct_year_month(rides, 2022, 1)

rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet')
